dir(recommendation-service): service dir 하위의 recommendation dir로 관련 서비스 이동 #26
