+++ src/main/java/com/my/firstbeat/web/service/recommemdation/RecommendationServiceWithRedis.java
@@ -1,6 +1,5 @@
-package com.my.firstbeat.web.service;
+package com.my.firstbeat.web.service.recommemdation;
 
-import com.github.benmanes.caffeine.cache.Cache;
 import com.my.firstbeat.client.spotify.SpotifyClient;
 import com.my.firstbeat.client.spotify.dto.response.RecommendationResponse;
 import com.my.firstbeat.client.spotify.ex.SpotifyApiException;
@@ -13,13 +12,20 @@ import com.my.firstbeat.web.domain.track.TrackRepository;
 import com.my.firstbeat.web.domain.user.User;
 import com.my.firstbeat.web.ex.BusinessException;
 import com.my.firstbeat.web.ex.ErrorCode;
+import com.my.firstbeat.web.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.redis.core.ListOperations;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import java.util.*;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Queue;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.TimeUnit;
@@ -27,36 +33,36 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.stream.Collectors;
 
-
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Slf4j
-//락 적용
-public class RecommendationService {
+public class RecommendationServiceWithRedis {
 
     private final SpotifyClient spotifyClient;
     private final UserService userService;
     private final GenreRepository genreRepository;
     private final PlaylistRepository playlistRepository;
     private final TrackRepository trackRepository;
+    private final RedisTemplate<String, Object> redisTemplate;
 
-    private final Cache<Long, Queue<TrackRecommendationResponse>> recommendationsCache;
-    private final Map<Long, ReentrantLock> userLocks = new ConcurrentHashMap<>(); //유저 별 락
-    private static final int REFRESH_THRESHOLD = 5; // 5개 남으면 새로 다시 요청
-    private static final int RECOMMENDATIONS_SIZE = 20; //한 번에 받아오는 추천 트랙 수
-    private static final int MAX_ATTEMPTS = 20; //추천한 곡이 이미 유저의 플레이리스트에 있는 경우 다시 추천 큐에서 꺼내올 수 있는 최대 횟수
+    private final Map<Long, ReentrantLock> userLocks = new ConcurrentHashMap<>(); // 유저별 락
+
+    private static final String RECOMMENDATION_KEY_PREFIX = "recommendation:user:";
+    private static final int REFRESH_THRESHOLD = 5;
+    private static final int RECOMMENDATIONS_SIZE = 20;
+    private static final int MAX_ATTEMPTS = 20;
     private static final int SEED_MAX = 5;
 
-
+    private static final long CACHE_TTL_HOURS = 24;
 
     public TrackRecommendationResponse getRecommendations(Long userId) {
         User user = userService.findByIdOrFail(userId);
-        Queue<TrackRecommendationResponse> recommendations =
-                recommendationsCache.get(userId, key -> new ConcurrentLinkedQueue<>());
+        String redisKey = RECOMMENDATION_KEY_PREFIX + userId;
+
 
         //락 획득 전 반환할게 있다면 빠른 반환 시도
-        TrackRecommendationResponse quickTry = recommendations.poll();
+        TrackRecommendationResponse quickTry = popRecommendation(redisKey);
         if(quickTry != null && !trackRepository.existsInUserPlaylist(user, quickTry.getSpotifyTrackId())){
             return quickTry;
         }
@@ -68,15 +74,13 @@ public class RecommendationService {
         for(int attempts = 1; attempts <= MAX_ATTEMPTS; attempts++) {
             userLock.lock();
             try {
-                if(needsRefresh(recommendations)) {
-                    refreshRecommendations(user);
-                    //갱신 후 캐시에서 추천 리스트 다시 가져옴
-                    recommendations = recommendationsCache.get(userId, key -> new ConcurrentLinkedQueue<>());
+                if(needsRefresh(redisKey)) {
+                    refreshRecommendations(user, redisKey);
                 }
             } finally {
                 userLock.unlock();
             }
-            TrackRecommendationResponse recommendation = recommendations.poll();
+            TrackRecommendationResponse recommendation = popRecommendation(redisKey);
             if(recommendation == null) {
                 throw new BusinessException(ErrorCode.FAIL_TO_GET_RECOMMENDATION);
             }
@@ -89,7 +93,11 @@ public class RecommendationService {
         throw new BusinessException(ErrorCode.MAX_RECOMMENDATION_ATTEMPTS_EXCEED);
     }
 
-    private void refreshRecommendations(User user) {
+    private TrackRecommendationResponse popRecommendation(String redisKey) {
+        return (TrackRecommendationResponse) redisTemplate.opsForList().leftPop(redisKey);
+    }
+
+    private void refreshRecommendations(User user, String redisKey) {
         try {
             String seedGenres = getSeedGenres(user);
             String seedTracks = getSeedTracks(user);
@@ -97,13 +105,15 @@ public class RecommendationService {
             RecommendationResponse spotifyRecommendations =
                     spotifyClient.getRecommendations(seedTracks, seedGenres, RECOMMENDATIONS_SIZE);
 
-            Queue<TrackRecommendationResponse> recommendations = recommendationsCache.get(user.getId(), key -> new ConcurrentLinkedQueue<>());
+            ListOperations<String, Object> listOps = redisTemplate.opsForList();
+
             spotifyRecommendations.getTracks()
                     .stream()
                     .map(TrackRecommendationResponse::new)
-                    .forEach(recommendations::offer);
-            // 데이터 갱신
-            recommendationsCache.put(user.getId(), recommendations);
+                    .forEach(track -> listOps.rightPush(redisKey, track));
+
+            //24시간 후 만료
+            redisTemplate.expire(redisKey, CACHE_TTL_HOURS, TimeUnit.HOURS);
         } catch (SpotifyApiException e){
             log.error("Spotify API 호출 실패 - 유저 ID: {}, 원인: {}", user.getId(), e.getMessage(), e);
             throw e;
@@ -120,24 +130,28 @@ public class RecommendationService {
         AtomicInteger successCount = new AtomicInteger(0);
         AtomicInteger failCount = new AtomicInteger(0);
 
-        var refreshEntries = recommendationsCache.asMap()
-                .entrySet()
-                .stream()
-                .filter(entry -> needsRefresh(entry.getValue()))
-                .toList();
-        log.info("추천 트랙 갱신 백그라운드 리프레시 시작 - 대상 유저 수: {}", refreshEntries.size());
+        Set<String> keys = redisTemplate.keys(RECOMMENDATION_KEY_PREFIX + "*");
+        if(keys.isEmpty()){
+            return;
+        }
 
-        refreshEntries
-                .forEach(entry -> {
-                    Long userId = entry.getKey();
-                    ReentrantLock userLock = userLocks.computeIfAbsent(userId, key -> new ReentrantLock());
+        List<String> refreshKeys = keys.stream()
+                .filter(this::needsRefresh)
+                .toList();
+
+        log.info("추천 트랙 갱신 백그라운드 리프레시 시작 - 대상 유저 수: {}", refreshKeys.size());
+
+        refreshKeys
+                .forEach(key -> {
+                    Long userId = Long.parseLong(key.replace(RECOMMENDATION_KEY_PREFIX, ""));
+                    ReentrantLock userLock = userLocks.computeIfAbsent(userId, k -> new ReentrantLock());
                     boolean locked = false;
 
                     try{
                         locked = userLock.tryLock(1000, TimeUnit.MILLISECONDS);
                         if(locked){
                             User user = userService.findByIdOrFail(userId);
-                            refreshRecommendations(user);
+                            refreshRecommendations(user, key);
                             successCount.incrementAndGet();
                         }
                     } catch (Exception e){
@@ -156,11 +170,13 @@ public class RecommendationService {
     //userLock 클린업
     @Scheduled(fixedRate = 1, timeUnit = TimeUnit.HOURS)
     private void cleanupUserLocks(){
-        userLocks.keySet().removeIf(userId -> !recommendationsCache.asMap().containsKey(userId));
+        Set<String> keys = redisTemplate.keys(RECOMMENDATION_KEY_PREFIX + "*");
+        userLocks.keySet().removeIf(userId -> !keys.contains(RECOMMENDATION_KEY_PREFIX + userId));
     }
 
-    public boolean needsRefresh(Queue<TrackRecommendationResponse> recommendations) {
-        return recommendations.size() <= REFRESH_THRESHOLD;
+    public boolean needsRefresh(String redisKey) {
+        Long size = redisTemplate.opsForList().size(redisKey);
+        return size == null || size <= REFRESH_THRESHOLD;
     }
