+++ src/main/java/com/my/firstbeat/web/service/recommemdation/RecommendationServiceWithoutLock.java
@@ -1,4 +1,4 @@
-package com.my.firstbeat.web.service;
+package com.my.firstbeat.web.service.recommemdation;
 
 import com.my.firstbeat.client.spotify.SpotifyClient;
 import com.my.firstbeat.client.spotify.dto.response.RecommendationResponse;
@@ -11,6 +11,7 @@ import com.my.firstbeat.web.domain.track.TrackRepository;
 import com.my.firstbeat.web.domain.user.User;
 import com.my.firstbeat.web.ex.BusinessException;
 import com.my.firstbeat.web.ex.ErrorCode;
+import com.my.firstbeat.web.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
