+++ src/test/java/com/my/firstbeat/web/service/RecommendationServiceTest.java
@@ -11,6 +11,7 @@ import com.my.firstbeat.web.domain.track.TrackRepository;
 import com.my.firstbeat.web.domain.user.Role;
 import com.my.firstbeat.web.domain.user.User;
 import com.my.firstbeat.web.dummy.DummyObject;
+import com.my.firstbeat.web.service.recommemdation.RecommendationService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
