+++ src/test/java/com/my/firstbeat/web/service/RecommendationServiceWithoutLockTest.java
@@ -12,6 +12,7 @@ import com.my.firstbeat.web.domain.track.TrackRepository;
 import com.my.firstbeat.web.domain.user.Role;
 import com.my.firstbeat.web.domain.user.User;
 import com.my.firstbeat.web.dummy.DummyObject;
+import com.my.firstbeat.web.service.recommemdation.RecommendationServiceWithoutLock;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
